Update patch set 1
Patch Set 1:

> While it seems a little shaky to leave this in long-term (it
 > assumes that the other repo will be locally cloned already),

No it does not assume the repo will be locally cloned. It will clone it remotely. What is nice about using relative paths is for example:

acumos-python-client: ssh://acumos-jobbuilder@gerrit.acumos.org:29418/acumos-python-client
model-schema: ../model-schema

Git Submodules will translate model-schema to:

 ssh://acumos-jobbuilder@gerrit.acumos.org:29418/model-schema


So it will cause git-submodule command to reconstruct the full URL based on the URL of the parent and then clone.

This is a good pattern to use especially if projects get moved around. This means if someone clones these 2 same projects on github or a local mirror. The translated full URL for the "model-schema" project will always be the same URL used by the parent project. This is actually a very handy way to clone submodules if you expect to use mirrors or clone from alternative urls.

Patch-set: 1