Update patch set 1
Patch Set 1: Verified-1 Code-Review-1

> I see this line that's still in setup.py, is it not effective?
 > 
 > python_requires='>=3.4',

We've had issues with people using Python 2 and still being able to install, despite this line. It may be that setup.py was invoked directly, or perhaps an old version of pip was being used that did not respect python_requires.

Patch-set: 1
Label: Code-Review=-1
Label: Verified=-1