Update patch set 2
Patch Set 2:

Hi Paul,

We've deployed a new docker image for the VM that now contains pyenv in the /opt/pyenv directory and preinstalled python 3.4, 3.5, and 3.6 for use. Additionally protoc is now available in /usr/local/bin/protoc.

The job appears to pick up these tools now but is still failing. Can you investigate and let us know if the remaining issue requires changes in the image or not? We can finalize the deployed image once you confirm that it has all the tooling you need. Detailed logs are available in the logs server link in the Jenkins comment above.

Hope this helps,
Thanh

Patch-set: 2