Update patch set 1
Patch Set 1:

The footer is the block of lines with change-id and signed-of-by.  Right now the Gerrit system is picky about the order of lines in footer.  Put issue-id line either in the body of the message, or as the first line of the footer, and you should be fine.

Patch-set: 1