Update patch set 4
Patch Set 4:

Aimee, I think these changes are mostly good. I see one problem, which is that you've removed "Acumos" from the title of the readme and replaced it with "Python". In fact it seems that you use the "Python ___" prefix in each doc title.

When the documentation is visible on PyPI (https://pypi.org/project/acumos/) users already know that this is a Python package, because this is the Python package index. IMO it is likely that Python users will see the PyPI documentation before they see documentation on the Acumos website, as Google tends to point folks to PyPI for Python projects.

I think that the RST doc titles should be "Acumos Python ___" so that readers on PyPI understand the context that this is an Acumos project. Otherwise "Python ___" is a bit redundant, given it's a Python package index.

Patch-set: 4
Label: Verified=0