Update patch set 2
Patch Set 2:

> Where will this report be utilized? Was this a technical debt/build
 > requirement?  Don't forget to tie to an issue number, if available.

The report isn't consumed by any downstream processes (yet), and instead is just logged in the build console.

This was tech debt I've been meaning to take care of since we were using BitBucket. I don't have an issue ID for this. I suppose I could create a jira issue now, but I find it odd to open and immediately close an issue.

I've been manually tracking my coverage, but decided to get this over with so that this recipe could be reused in the upcoming model runner repo.

Patch-set: 2
Label: Code-Review=0