Update patch set 1
Patch Set 1:

> @Paul, Because it will be asked, could you also include a
 > documented example of including a binary asset in the package;
 > perhaps a model, an image, or a simple binary CSV file for testing

Hi Eric,

That feature isn't a part of this PR, and more generally isn't a feature of this library yet.

I'm not sure what you have in mind for including images / CSV files with the model. Would the model need this image, i.e. should it be included in the Docker image? Or is it for documentation?

Patch-set: 1