Update patch set 1
Patch Set 1:

> Approved, but why the extra footwork in munging the JWT access?

Namely to hide what the token was being called in the config. Multiple spots referred to the jwt by a string literal or kwarg. Since it was contained in the same module I didn't mind but now that a separate module sets the jwt for a test case, I added formal mutators.

Patch-set: 1