Update patch set 1
Patch Set 1:

The first sentence was : acumos is a client library. It's a little bit disturbing for a new comer because acumos is not a client library, acumos it's a linux foundation project. Perhaps we have to explain clearly that the on-boarding python client is called "acumos".

Patch-set: 1
Reviewer: Gerrit User 212 <212@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=0