Update patch set 23
Patch Set 23:

> Build Failed
 > 
 > https://jenkins.acumos.org/job/acumos-python-client-tox-verify-master/183/
 > : FAILURE
 > 
 > Logs: https://logs.acumos.org/production/vex-yul-acumos-jenkins-1/acumos-python-client-tox-verify-master/183

The good news is that pinning pandas is allowing the unit tests to be ran. However, a unit test appears to be failing this review, specifically this line from the Jenkins log:

validate(metadata, schema)

Since the support of new media types has impacted the model metadata schema, a new metadata schema needs to be drafted, agreed upon by the Acumos community, and then version controlled in the model schema repo:

https://gerrit.acumos.org/r/gitweb?p=model-schema.git;a=tree

Then the acumos library needs to specify which schema version it is producing instances of:

https://github.com/acumos/acumos-python-client/blob/master/acumos/metadata.py#L33

And then the unit test will load the appropriate schema and validate what the library generates.

Patch-set: 23