Update patch set 1
Patch Set 1:

Copying my comment from https://gerrit.acumos.org/r/#/c/acumos-python-client/+/5143/ :

Looking at the Jenkins logs, it seems that the tensorflow backend dependency of the keras model is not being captured correctly. I'm not yet sure if this is due to a change in keras and/or tensorflow.

As a temporary solution, I would pin tensorflow and keras to a previously known working version, and we can work on this regression after your media type feature is working and merged.

Judging from this previous successful build:

https://jenkins.acumos.org/job/acumos-python-client-tox-verify-master/152/console

I'd pin:

Keras==2.2.4
tensorflow==1.13.1

in testing/tox-requirements.txt

Patch-set: 1
Label: Verified=0