Update patch set 2
Patch Set 2:

Can any other context be added for a BZ2 limitation?
Also, if it's just pandas that's effected, can we lift the version pegs for keras and tensorflow.  Those two efforts are much faster moving (versus the speed of pandas and our updates to this library) and if it's only something to do with pandas, the dependency resolver will properly enforce that limitation.

Patch-set: 2