Update patch set 3
Patch Set 3:

Sorry for the noise Philippe. Indeed, there seems to be a problem with more recent versions of either keras or tensorflow, in which the acumos library fails to pick up the dependencies correctly.

So it would probably be best to keep the version pins in keras and tensorflow, and merge this patch so that at least the acumos library can be tested without failing.

Patch-set: 3
Reviewer: Gerrit User 16 <16@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=0