Update patch set 27
Patch Set 27:

Regarding unit tests, I think we need a test that captures end-to-end functionality.

For example, we need to check that the model method can be invoked correctly, and that the generated metadata is correct.

What I might do is, in a for-loop, iterate over each raw type we support and:

1. Define a dummy model
2. Dump the model
3. Check the media type in the model metadata is correct
4. Load the dumped model back
5. Invoke the wrapped model's appropriate I/O methods, e.g. from_dict, to_dict

Here's an example of such a test: https://github.com/acumos/acumos-python-client/blob/master/acumos/tests/test_wrapped.py#L257

You need not use this test; there could be a specific test for raw types.

Patch-set: 27