Update patch set 1
Patch Set 1:

Hi Philippe,

I'm curious where this request to store this image in this repository is coming from. Does every repository need to version control this image? Would it be possible to have the logo stored in 1 repository, and then we can simply reference the image URL in our readme.rst?

Patch-set: 1