Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi Philippe,
> 
> I'm curious where this request to store this image in this repository is coming from. Does every repository need to version control this image? Would it be possible to have the logo stored in 1 repository, and then we can simply reference the image URL in our readme.rst?

Hi Paul, the need is to have the logo displayed in Github (it's a legal need). I guess that all the component name in which "acumos" appeared must have this acumos logo. for example : https://github.com/acumos/acumos-java-client. So the place to store this image is not realy important I put it in docs/images because it make sens to store images here. When you said "Would it be possible to have the logo stored in 1 repository" you meant in 1 repository outside the acumos-python-client folder right ? So I'm not agains't that but I don't know where we can stored this image outside acumos-python-client

Patch-set: 1