Update patch set 27
Patch Set 27:

I don't believe this should be merged as-is, because we don't have an end-to-end unit test which dumps a model, loads it back, and uses the "raw" I/O APIs of the model wrapper. See my last comment. 

Without this type of unit test, we have no evidence that dumped models will be loaded and used as expected with the model runner.

Patch-set: 27