Update patch set 27
Patch Set 27: -Code-Review

It seems to me that we can go even further in end-to-end testing, if we wanted to; for example, we don't have a test that packages a model as a microservice, then deploys the microservice and feeds it input and checks the responses--and this would be a true end-to-end test of how the functionality works in Acumos.

So it's a question of what level of end-to-end testing we want to build in here, isn't it?

Patch-set: 27
Reviewer: Gerrit User 8 <8@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=0