Update patch set 27
Patch Set 27:

> It seems to me that we can go even further in end-to-end testing,
 > if we wanted to; for example, we don't have a test that packages a
 > model as a microservice, then deploys the microservice and feeds it
 > input and checks the responses--and this would be a true end-to-end
 > test of how the functionality works in Acumos.
 > 
 > So it's a question of what level of end-to-end testing we want to
 > build in here, isn't it?

The standalone and DCAE model runners already have unit tests at their respective "microservice" interfaces.

The client library provides a wrapped model API to both of those model runners, so the model wrapper API needs to be tested here in this repo. IMO the model wrapper should probably be its own independent Python package.

Indeed, Acumos (at least the microservice generator) should be doing a full end-to-end test of build-deploy-test for baseline models for each language. This would mainly be to test that the docker image is built correctly. I believe we've discussed this in the past.

Patch-set: 27
Label: Code-Review=0