Update patch set 1
Patch Set 1:

Thanks paul. The merge fails because test.pypi.org does not allow overwrite of a package; if it does, I don't know how to enable/force it, so the template doesn't attempt. This also means this version cannot be promoted to release bcos the real pypi.org has the same rule.

Philippe when you bump the version, then a merge will succeed, and a release will be possible.

At least now the README looks better :)

Patch-set: 1