Update patch set 7
Patch Set 7:

This challenge of including files from docs/ in the Python source distribution is much, much more important than I first realized.  

Unlike a Java project, where RST files are only for ReadTheDocs, here the RST files are part of the package that is pushed to PyPI. If those files change, then the package must be rebuilt!  So that's a key assumption of mine that I now see was/is false.

The next challenge is that the Python setup tools are written to work only with content within the project root, which is the directory where the setup.py and other files live.  Previously that was the git repo root; now it's the directory "acumos-package".  However, this reorg puts the docs/ directory outside that project root, effectively making the files in visible to the Python packager tools.

We probably have to abandon this and go back to the old organization.  Thank you Philippe for trying and I apologize for costing you so much time here.

Patch-set: 7