Update patch set 4
Patch Set 4:

Paul, this change set corrects a botch of #6275.  While we were reorganizing the files, some packages changed by dropping support for Python 3.5, so we also dropped support for Python 3.5.  The botch was that we didn't increment the version string from 0.9.1, so the push to test.pypi.org failed.   By merging this, you will stage a release of 0.9.2 at test.pypi.org that properly advertises what it supports, which sets Philippe up to release it to pypi.org.

HTH

Patch-set: 4