Update patch set 10
Patch Set 10: Code-Review-1

This change set alters the code behavior but does not increment the version.  When you change the version, you must decide if you are working on a development/test version, and use a suffix like "0.9.3a1", or if you are ready to release the next version and don't use a dev/test suffix.  This is standardized in the Python community, for full details: https://www.python.org/dev/peps/pep-0440/

Patch-set: 10
Label: Code-Review=-1