Update patch set 22
Patch Set 22:

> Patch Set 22: Verified+1
> 
> Build Successful 
> 
> https://jenkins.acumos.org/job/acumos-python-client-pypi-verify-master/106/ : SUCCESS
> 
> Logs: https://logs.acumos.org/production/vex-yul-acumos-jenkins-1/acumos-python-client-pypi-verify-master/106
> 
> https://jenkins.acumos.org/job/documentation-rtd-verify-master/5264/ : SUCCESS
> 
> Logs: https://logs.acumos.org/production/vex-yul-acumos-jenkins-1/documentation-rtd-verify-master/5264

Ok, I think I understood. For running the test, a mock server is created to emulate Acumos, so the API must be defined somewhere in the code. I have to modify this API in the mock server to take into account this new API parameter called "dockerImageUri". I thought that an Acumos server was hosted in the LF environement that is why I created the LF ticket

So it will take time for me to modify the "mock" API, what I can propose you is to postpone this mock API modification later in another change. I remove session.py for which the modification only invoke the mock API, and I will create another change later.

I would like to merge this change quickly because it brings some modification related to the defect on unstructured type, it's an urgent matter thanks.

please give +1

Patch-set: 22
Label: Code-Review=0