Update patch set 1
Patch Set 1: Code-Review+2

(2 comments)

Although this targets a JIRA issue, we need to make sure it's inline with the schema definitions, which it appears it is.  These requests kept falling below the fold and I'm not good at inbox-zero.  Thanks for the persistent and clean implementations for new pythonic tests.

Patch-set: 1
Reviewer: Gerrit User 10 <10@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=+2