Zip the model bundle.
Also add an optional flag to replace and existing model
when dumping instead of raising an exception.

First pass on end to end tests, for now we check if a
model can be pushed on a real acumos instance
with a matrix of input and output types.
We also make sure that models given as examples
can be pushed.

This merge is still WIP, I'd like to add real end-to-end
tests : push a model, get the docker image, run it and
test it.

Issue-id: ACUMOS-4274
Issue-id: ACUMOS-4273
Signed-off-by: Jonathan GAYVALLET <jonathan.gayvallet@orange.com>
Change-Id: Iee50ff868b60d33be21bd1240effc65392b562ff
Signed-off-by: Jonathan GAYVALLET <jonathan.gayvallet@orange.com>