Zip the model bundle.
Added an optional flag to replace and existing model
when dumping instead of raising an exception.

Fixed h5py version in tox-requirements as h5py 3 breaks with our
pinned version of keras.

Session.push now returns the URI to the docker image as an unicode string
in addition to writing it to the console logs.

Issue-id: ACUMOS-4273
Signed-off-by: Jonathan GAYVALLET <jonathan.gayvallet@orange.com>
Change-Id: Iee50ff868b60d33be21bd1240effc65392b562ff